[ refactor ] FolderPage api 및 eventHandler 로직 hook으로 관리
