+++ src/pages/FolderPage.jsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 
 import Hero from '../components/Hero/Hero';
@@ -8,64 +7,12 @@ import Search from '../components/Contents/CardSearchBar/CardSearchBar';
 import CardList from '../components/Contents/CardList/CardList';
 import FolderCollection from '../components/Contents/FolderCollection/FolderCollection';
 
-import getFetch from '../utils/getFetch';
-import getFormattedCamelCaseData from '../utils/getFormattedCamelCaseData';
 import Modal from '../components/Modal/Modal';
-import { DEFALUT_MODAL_VALUE } from '../Constants/Constants';
+import { useFolder, useModal } from './useFolderPage.hook';
 
 const FolderPage = ({ userData }) => {
-  const [folderData, setFolderData] = useState([]);
-  const [folderCardData, setFolderCardData] = useState([]);
-
-  // Modal
-  const [modal, setModal] = useState(DEFALUT_MODAL_VALUE);
-
-  const showModal = (modalValue) => {
-    setModal(modalValue);
-  };
-
-  const closeModal = () => {
-    setModal(DEFALUT_MODAL_VALUE);
-  };
-
-  // 폴더의 전체 버튼을 클릭했을 때 가져올 데이터
-  const handleOverviewCardButtonClick = () => {
-    try {
-      getFetch('bootcamp-api.codeit.kr', 'api/users/1/links').then((cardData) => {
-        setFolderCardData(() => {
-          return getFormattedCamelCaseData(cardData.data);
-        });
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  // 폴더의 전체 버튼이 아닌 버튼을 클릭했을 때 가져올 데이터
-  const handleFilteredCardButtonClick = (id) => {
-    try {
-      getFetch('bootcamp-api.codeit.kr', `api/users/1/links?folderId=${id}`).then((cardData) => {
-        setFolderCardData(() => {
-          return getFormattedCamelCaseData(cardData.data);
-        });
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  // 폴더 버튼들을 가지고 있는 데이터
-  useEffect(() => {
-    try {
-      getFetch('bootcamp-api.codeit.kr', 'api/users/1/folders').then((FolderData) => {
-        setFolderData(() => {
-          return getFormattedCamelCaseData(FolderData.data);
-        });
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  }, []);
+  const { modal, setModal, showModal, closeModal } = useModal();
+  const { folderData, folderCardData, handleOverviewCardButtonClick, handleFilteredCardButtonClick } = useFolder();
 
   return (
     <FolderPageWrapper>
